+++ tools/run-np-subsystem-v0.5.184.js
@@ -1,3 +1,0 @@
-#!/usr/bin/env node
-"use strict";
-require("../tests/run-np-subsystem.js");