+++ tools/run-regression-suite.js
@@ -1,3 +1,0 @@
-#!/usr/bin/env node
-"use strict";
-require("../tests/run-regression.js");